Create Exercise-5-HomeSales
Danielle, Edward, and Francis are three salespeople at Holiday Homes. Write an
application named HomeSales that prompts the user for a salesperson initial (D,
E, or F). Either uppercase or lowercase initials are valid. While the user does
not type Z, continue by prompting for the amount of a sale. Issue an error
message for any invalid initials entered. Keep a running total of the amounts
sold by each salesperson. After the user types Z or z for an initial, display
each salesperson’s total, a grand total for all sales, and the name of the
salesperson with the highest total.