Create Exercise-15-GuessingGame3
In Chapter 4, you created a program that generates a random number, allows a
user to guess it, and displays a message indicating whether the guess is too
low, too high, or correct. Now, create a modified program called GuessingGame2
in which the user can continue to enter values until the correct guess is made.
After the user guesses correctly, display the number of guesses made.